#GH-3 added permission check in get and set config methods
